[STYLE] QuizChoiceView 파일 분리 (#24)
